+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/items/pouch/PouchTooltip.java
@@ -1,0 +1,8 @@
+package insane96mcp.insanesurvivaloverhaul.module.items.pouch;
+
+import net.minecraft.core.NonNullList;
+import net.minecraft.world.inventory.tooltip.TooltipComponent;
+import net.minecraft.world.item.ItemStack;
+
+public record PouchTooltip(NonNullList<ItemStack> items) implements TooltipComponent {
+}